requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
anyio==3.6.2
    # via httpcore
azqueuetweeter==0.1.1
    # via -r requirements.in
azure-core==1.26.1
    # via
    #   azure-storage-queue
    #   msrest
azure-storage-queue==12.5.0
    # via azqueuetweeter
certifi==2022.9.24
    # via
    #   httpcore
    #   httpx
    #   msrest
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via typer
cryptography==38.0.3
    # via azure-storage-queue
gh-issues==0.0.3.dev16
    # via -r requirements.in
h11==0.12.0
    # via httpcore
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via gh-issues
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
isodate==0.6.1
    # via msrest
markdown-it-py==2.1.0
    # via gh-issues
mdurl==0.1.2
    # via markdown-it-py
msrest==0.7.1
    # via azure-storage-queue
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   tweepy
pycparser==2.21
    # via cffi
python-dotenv==0.21.0
    # via -r requirements.in
requests==2.28.1
    # via
    #   azure-core
    #   msrest
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==1.3.1
    # via
    #   msrest
    #   tweepy
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via
    #   azure-core
    #   isodate
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tweepy==4.12.1
    # via azqueuetweeter
typer==0.7.0
    # via -r requirements.in
typing-extensions==4.4.0
    # via azure-core
urllib3==1.26.12
    # via requests
